ml_server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
boto==2.49.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
contourpy==1.3.2
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
Cython==3.1.0
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
docker-pycreds==0.4.0
exceptiongroup==1.3.0
executing==2.2.0
fasteners==0.19
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gcs-oauth2-boto-plugin==3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-reauth==0.1.1
googleapis-common-protos==1.70.0
gsutil==5.34
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
-e git+https://github.com/learnables/learn2learn.git@db5cee3795490071282bbdfa81179ef732349196#egg=learn2learn
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauth2client==4.1.3
ollama==0.4.8
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
playsound==1.3.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyOpenSSL==24.2.1
pyparsing==3.2.3
pyright==1.1.400
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyu2f==0.1.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retry-decorator==1.1.1
rsa==4.7.2
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
